refactor: remove store usage from CostByRegionWidget
